requirements.txt
duckdb==1.1.3
pandas==2.2.3
streamlit==1.39.0
plotly==5.24.1
anthropic==0.39.0
python-dotenv==1.0.1
